pyton/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
openai==1.54.4
pymupdf==1.24.12
python-docx==1.1.2
python-multipart==0.0.12
slowapi==0.1.9
aiofiles==24.1.0
